requirements.txt
streamlit>=1.25.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.40.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
numpy
